feat: only trigger emit on continue
